+++ admin/taikhoan/delete.php
@@ -22,6 +22,7 @@
                                     <th>Vai trò</th>
                                 </tr>
                             </thead>
+                            <tbody>
                             <?php
                             if (!empty($listtk)) {
                                 // Hiển thị thông tin danh mục
@@ -37,7 +38,7 @@
                                             <td>'.($role == 1 ? 'Admin' : 'User').'</td>
                                             <td>
                                                 <a href="index.php?act=restoretk&id=' . $id . '"><button class="btn btn-primary">Khôi phục</button></a>
-                                                <a href="index.php?act=deletetk&id=' . $id . '"><button class="btn btn-danger">Xóa</button></a>
+                                                <a href="index.php?act=deletetk&id=' . $id . '"><button onclick="confirmDelete(' . $id . ')" class="btn btn-danger">Xóa</button></a>
                                             </td>
                                         </tr>';
                                 }
@@ -45,22 +46,7 @@
                                 echo '<tr><td colspan="4">Không có tài khoản nào.</td></tr>';
                             }
                             ?>
-                            <tbody>
-                                <!-- <tr>
-                                        <td><input type="checkbox"></td>
-                                        <td>1</td>
-                                        <td>text</td>
-                                        <td>text</td>
-                                        <td>no photo</td>
-                                        <td>
-                                            <a href="index.php?act=updatebn"><button class="btn btn-primary">Sửa</button></a>
-                                            <a href="#"><button class="btn btn-danger">Xóa</button></a>
-
-                                        </td>
-                                    </tr> -->
                             </tbody>
-
-
                         </table>
                         <ul class="pagination">
                             <?php
@@ -82,5 +68,18 @@
         <!-- /.row -->
     </div>
     <!-- /.container-fluid -->
+    <script>
+    // Hàm để hiển thị hộp thoại xác nhận xóa
+    function confirmDelete(id) {
+        var confirmed = confirm("Bạn có chắc chắn muốn xóa không?");
+        if (confirmed) {
+            // Nếu người dùng chọn Yes, chuyển hướng đến trang xóa với tham số id
+            window.location.href = "index.php?act=deletetk&id=" + id;
+        } else {
+            event.preventDefault();
+            // Nếu người dùng chọn No, không thực hiện hành động gì
+        }
+    }
+</script>
 </section>
 <!-- /.content -->
